fix(balancer) delay initialization of balancer objects
When upstreams include targets using hostnames instead of IPs, the
initialization of balancer objects needs to perform DNS resolution, which
involves opening a socket, which is not available in the init_worker phase.

This patch delays balancer initialization to the access phase, using
a lock to ensure that initialization happens only once: the initialization
possibly contains operations that yield (e.g. DNS resolutions), so the
same worker may start servicing another access phase at the same time.
The use of the lock prevents that. The sentinel variables in both
`kong.core.balancer` and `kong.core.handler` ensure that this is a
one-time cost at startup.